+++ atcoder-problems-backend/src/server/ranking/mod.rs
@@ -3,7 +3,6 @@ use crate::server::AppData;
 use actix_web::{web, HttpRequest, HttpResponse, Result};
 use async_trait::async_trait;
 use serde::{de::DeserializeOwned, Deserialize, Serialize};
-use sql_client::models::UserSum;
 use std::ops::Range;
 
 pub(crate) mod ac_count;
@@ -77,7 +76,6 @@ pub(crate) trait UserRankSelector<A: Sync + Send + Clone + 'static> {
     }
 }
 
-// ranking requests
 #[derive(Deserialize)]
 pub(crate) struct RankingRequest {
     from: usize,
@@ -90,20 +88,6 @@ impl RankingRequestFormat for RankingRequest {
     }
 }
 
-#[derive(Deserialize)]
-pub(crate) struct LanguageRankingRequest {
-    from: usize,
-    to: usize,
-    language: String,
-}
-
-impl RankingRequestFormat for LanguageRankingRequest {
-    fn range(&self) -> Range<usize> {
-        (self.from)..(self.to)
-    }
-}
-
-// ranking responses
 #[derive(Serialize)]
 pub(crate) struct RankingResponse {
     user_id: String,
@@ -112,9 +96,6 @@ pub(crate) struct RankingResponse {
 
 impl RankingResponseFormat for RankingResponse {}
 
-impl RankingResponseFormat for UserSum {}
-
-// user rank requests
 #[derive(Deserialize)]
 pub(crate) struct UserRankRequest {
     user: String,
@@ -122,7 +103,6 @@ pub(crate) struct UserRankRequest {
 
 impl UserRankRequestFormat for UserRankRequest {}
 
-// user rank responses
 #[derive(Serialize)]
 pub(crate) struct UserRankResponse {
     count: i64,
@@ -130,12 +110,3 @@ pub(crate) struct UserRankResponse {
 }
 
 impl UserRankResponseFormat for UserRankResponse {}
-
-#[derive(Debug, Serialize)]
-pub(crate) struct LanguageUserRankResponse {
-    language: String,
-    count: i64,
-    rank: i64,
-}
-
-impl UserRankResponseFormat for LanguageUserRankResponse {}
