+++ atcoder-problems-backend/tests/test_server_e2e_rated_point_sum_ranking.rs
@@ -61,9 +61,9 @@ async fn test_rated_point_sum_ranking() {
     assert_eq!(
         response,
         json!([
-            {"user_id":"u2","point_sum":2},
-            {"user_id":"u1","point_sum":1},
-            {"user_id":"u3","point_sum":1}
+            {"user_id":"u2","point_sum":2,"count":2},
+            {"user_id":"u1","point_sum":1,"count":1},
+            {"user_id":"u3","point_sum":1,"count":1}
         ])
     );
 
@@ -79,8 +79,8 @@ async fn test_rated_point_sum_ranking() {
     assert_eq!(
         response,
         json!([
-            {"user_id":"u1","point_sum":1},
-            {"user_id":"u3","point_sum":1}
+            {"user_id":"u1","point_sum":1,"count":1},
+            {"user_id":"u3","point_sum":1,"count":1}
         ])
     );
     let response = reqwest::get(url(
@@ -95,7 +95,7 @@ async fn test_rated_point_sum_ranking() {
     assert_eq!(
         response,
         json!([
-            {"user_id":"u2","point_sum":2}
+            {"user_id":"u2","point_sum":2,"count":2}
         ])
     );
