+++ atcoder-problems-frontend/src/api/APIClient.ts
@@ -99,18 +99,8 @@ export const useUserStreakRank = (user: string) => {
 };
 
 export const useSumRanking = (from: number, to: number) => {
-  const fetcher = async (url: string) => {
-    const ranking = await fetchTypedArray<SumRankingEntry>(
-      url,
-      isSumRankingEntry
-    );
-    return ranking.map((entry) => ({
-      count: entry.point_sum,
-      user_id: entry.user_id,
-    }));
-  };
   const url = `${ATCODER_API_URL}/v3/rated_point_sum_ranking?from=${from}&to=${to}`;
-  return useSWRData(url, fetcher);
+  return useRankingV3(url);
 };
 
 export const useUserSumRank = (user: string) => {
